+++ plugin-build/modulegraph/src/main/kotlin/dev/iurysouza/modulegraph/graph/getProjectName.kt
@@ -1,4 +1,4 @@
-package dev.iurysouza.modulegraph.gradle.graph
+package dev.iurysouza.modulegraph.graph
 
 internal fun String.getProjectName(showFullPath: Boolean): String {
     return if (showFullPath) {
